Moved unit test files and profiling test decorator into test directory away from production code
