+++ monkey/tests/monkey_island/cc/conftest.py
@@ -1,3 +1,3 @@
 # Without these imports pytests can't use fixtures,
 # because they are not found
-from monkey_island.cc.test_common.fixtures import *  # noqa: F401,F403
+from tests.monkey_island.cc.mongomock_fixtures import *  # noqa: F401,F403
