+++ monkey/tests/monkey_island/cc/fixture_enum.py
