+++ monkey/tests/monkey_island/cc/models/zero_trust/test_scoutsuite_finding.py
@@ -1,5 +1,6 @@
 import pytest
 from mongoengine import ValidationError
+from tests.monkey_island.cc.fixture_enum import FixtureEnum
 
 import common.common_consts.zero_trust_consts as zero_trust_consts
 from monkey_island.cc.models.zero_trust.finding import Finding
@@ -7,7 +8,6 @@ from monkey_island.cc.models.zero_trust.monkey_finding_details import MonkeyFind
 from monkey_island.cc.models.zero_trust.scoutsuite_finding import ScoutSuiteFinding
 from monkey_island.cc.models.zero_trust.scoutsuite_finding_details import ScoutSuiteFindingDetails
 from monkey_island.cc.services.zero_trust.test_common.scoutsuite_finding_data import RULES
-from monkey_island.cc.test_common.fixtures import FixtureEnum
 
 MONKEY_FINDING_DETAIL_MOCK = MonkeyFindingDetails()
 MONKEY_FINDING_DETAIL_MOCK.events = ["mock1", "mock2"]
