+++ monkey/tests/monkey_island/cc/mongomock_fixtures.py
