+++ monkey/tests/monkey_island/cc/services/zero_trust/scoutsuite/test_scoutsuite_zt_finding_service.py
@@ -1,4 +1,5 @@
 import pytest
+from tests.monkey_island.cc.fixture_enum import FixtureEnum
 
 from monkey_island.cc.models.zero_trust.finding import Finding
 from monkey_island.cc.models.zero_trust.scoutsuite_finding import ScoutSuiteFinding
@@ -9,7 +10,6 @@ from monkey_island.cc.services.zero_trust.test_common.scoutsuite_finding_data im
     RULES,
     SCOUTSUITE_FINDINGS,
 )
-from monkey_island.cc.test_common.fixtures import FixtureEnum
 
 
 class TestScoutSuiteZTFindingService:
